Add page 400 Add page 403 Add page 404 Add Global Exception Handler - ApplicationExceptionHandler.java Add NullEntityException.java Add UserExistsException.java Add UserNotFoundException.java Edit SecurityConfig.java Add business-logic of Exception Handling in UserServiceImpl.java

+++ src/main/java/com/alexvar/springboot_rest/services/Impl/UserServiceImpl.java
@@ -1,5 +1,8 @@
 package com.alexvar.springboot_rest.services.Impl;
 
+import com.alexvar.springboot_rest.exception.NullEntityException;
+import com.alexvar.springboot_rest.exception.UserExistsException;
+import com.alexvar.springboot_rest.exception.UserNotFoundException;
 import com.alexvar.springboot_rest.model.Role;
 import com.alexvar.springboot_rest.model.User;
 import com.alexvar.springboot_rest.repositories.UserRepository;
@@ -7,8 +10,6 @@ import com.alexvar.springboot_rest.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 @Service
@@ -26,7 +27,9 @@ public class UserServiceImpl implements UserService{
     @Override
     public User create(User user) {
         if(userRepository.findUserByEmail(user.getEmail()).isPresent()) {
-            throw new EntityNotFoundException("User already exists!");
+            throw new UserExistsException("User already exists!");
+        }else if(user == null){
+            throw new NullEntityException("User is null!");
         }
             user.setPassword(passwordEncoder.encode(user.getPassword()));
             user.setRole(Role.USER);
@@ -35,11 +38,17 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public User readById(long id) {
+        if(!userRepository.existsById(id)){
+            throw new UserNotFoundException("User not found!");
+        }
         return userRepository.getReferenceById(id);
     }
 
     @Override
     public User update(User user) {
+        if(user == null){
+            throw new NullEntityException("User is null!");
+        }
         User newEmp = userRepository.findById(user.getId()).get();
         newEmp.setName(user.getName());
         newEmp.setSurname(user.getSurname());
@@ -50,11 +59,17 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public void delete(long id) {
+        if(!userRepository.existsById(id)){
+            throw new UserNotFoundException("User not found!");
+        }
         userRepository.deleteById(id);
     }
 
     @Override
     public List<User> getAll() {
+        if(userRepository.findAll().isEmpty()){
+            throw new UserNotFoundException("Users not found!");
+        }
         return userRepository.findAll();
     }
 }
